Phase 3: buildings, construction and population
- BuildingDef (ScriptableObject): footprint, cost, build time, population,
  drop-off flag, sprite.
- Building: construction site that villagers complete; blocks its footprint,
  contributes population and becomes a drop-off when finished.
- BuildingPlacer: grid placement with a ghost preview (green/red validity);
  press H to place a House, left-click to build, right-click/Esc to cancel.
- PlayerPopulation: cap from buildings; HUD shows current/cap.
- Villager: can now construct in addition to gathering.
- Selection: buildings and construction sites are clickable; SelectionInfoHud
  shows type, construction %, and role (drop-off / +population).
- Town Center and walls are now real Buildings (2x2 TC drop-off; 1x1 Walls).
- Gathering is integer-exact: ResourceNode amounts and villager loads are whole
  units, so resource totals add up exactly (no float truncation loss).
- Start with 150 wood / 50 food to allow construction immediately.
- GAME_IDEA.md: translated to English and lightly cleaned up.

+++ Assets/Scripts/Economy/ResourceNode.cs
@@ -11,14 +11,14 @@ namespace AditusBelli.Economy
     public class ResourceNode : MonoBehaviour
     {
         public ResourceType resourceType = ResourceType.Wood;
-        public float amount = 100f;
+        public int amount = 100;
 
         public static readonly List<ResourceNode> All = new();
 
         private Vector2Int _cell;
         private bool _blocked;
 
-        public bool IsDepleted => amount <= 0f;
+        public bool IsDepleted => amount <= 0;
 
         private void OnEnable() => All.Add(this);
         private void OnDisable() => All.Remove(this);
@@ -34,12 +34,12 @@ namespace AditusBelli.Economy
             _blocked = true;
         }
 
-        /// <summary>Removes up to 'requested' units and returns the amount actually taken.</summary>
-        public float Extract(float requested)
+        /// <summary>Removes up to 'requested' whole units and returns how many were actually taken.</summary>
+        public int Extract(int requested)
         {
-            float taken = Mathf.Min(requested, amount);
+            int taken = Mathf.Min(requested, amount);
             amount -= taken;
-            if (amount <= 0f) Deplete();
+            if (amount <= 0) Deplete();
             return taken;
         }
